requirements.txt
celery==5.2.3
Django==4.0.2
django-celery-results==2.2.0
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.13.1
ipython==8.0.1
json-enhanced==0.2.4
pandas==1.4.1
psycopg2==2.9.3
requests==2.27.1
